add txt file extension to sets file to silence parsing error
Even though we are reading file as text with ajax, which will specify
plain text on request header, the web server may specify something else
on the response header.  Whatever is running on biochemistry seems to
be responding with xml for data/sets.  So use .txt file extension to
give a hint to web server.