don't trigger change events when it doesn't actually change
There were a lot of 'false' change events when changing setup.  If
the new setup had the same ex path, or the same excitation or dye,
there was no check if it was the same and would trigger a change
event. Catch this as early as possible, and avoid triggering changes.