spekcheck-logo.svg: support for recent inkscape and rsvg-convert
Not sure what was wrong but while the previous svg would display correctly
on inkscape 0.48, it would fail on inkscape 0.92 and rsvg-convert tool.
Redid the svg template on inkscake 0.92.  Also changed to have link to the
visible spectrum instead of embebed it so we can even replace the bakground
spectrum via command line and have all in a Makefile.