Remove hints for non-existent dragging of filters from FilterStack.
We originally thought that a user could move filters out of the
FilterStacks (the excitation and emission path).  However, turns
out that it wasn't very intuitive and instead we added a close
button to have them removed instead.

This removes remains of that support across the whole path builder
interface, and limiting to the list of filters.